Fix Translation of Phetchabun / Phetchaburi
Notice the vowel differences. 
On `Line:1133` in `def get_provinces()`, this line attempts to match to Phetchabun province based on a partial name. The key difference between Phetchabun and Phetchaburi is the vowel is in long or short form. 
`Line:1133` is the short form, and should therefore map to Phetchaburi.

I'll attach some images to the pull request to help verify this.